getTime refactoring
