+++ src/client/clientlauncher.h
@@ -24,43 +24,6 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "client/inputhandler.h"
 #include "gameparams.h"
 
-// A small helper class
-class TimeGetter
-{
-public:
-	virtual u32 getTime(TimePrecision prec) = 0;
-};
-
-// A precise irrlicht one
-class IrrlichtTimeGetter: public TimeGetter
-{
-public:
-	IrrlichtTimeGetter(IrrlichtDevice *device):
-		m_device(device)
-	{}
-	u32 getTime(TimePrecision prec)
-	{
-		if (prec == PRECISION_MILLI) {
-			if (m_device == NULL)
-				return 0;
-			return m_device->getTimer()->getRealTime();
-		} else {
-			return porting::getTime(prec);
-		}
-	}
-private:
-	IrrlichtDevice *m_device;
-};
-// Not so precise one which works without irrlicht
-class SimpleTimeGetter: public TimeGetter
-{
-public:
-	u32 getTime(TimePrecision prec)
-	{
-		return porting::getTime(prec);
-	}
-};
-
 class ClientLauncher
 {
 public:
