+++ src/fontengine.cpp
@@ -17,6 +17,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
 #include "fontengine.h"
+#include "debug.h"
 #include "log.h"
 #include "config.h"
 #include "porting.h"
